add11

+++ src/services/contacts.js
@@ -45,33 +45,36 @@ export const getContactById = async (contactId, userId) => {
   return ContactsCollection.findOne({ _id: contactId, userId });
 };
 
-export const createContact = async (payload, userId) => {
-  return ContactsCollection.create(payload, userId);
+export const createContact = async (payload) => {
+  return ContactsCollection.create(payload);
 };
 
-export const updateContact = async (contactId, payload) => {
-  return ContactsCollection.findOneAndUpdate({ _id: contactId }, payload, {
-    new: true,
-  });
+export const updateContact = async (contactId, payload, userId) => {
+  return ContactsCollection.findOneAndUpdate(
+    { _id: contactId, userId },
+    payload,
+    {
+      new: true,
+    },
+  );
 };
 
-export const deleteContact = async (contactId) => {
-  return ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+  return ContactsCollection.findOneAndDelete({ _id: contactId, userId });
 };
 
-export const replaceContact = async (contactId, payload) => {
+export const replaceContact = async (contactId, payload, userId) => {
   const result = await ContactsCollection.findByIdAndUpdate(
-    contactId,
+    { _id: contactId, userId },
     payload,
     {
       new: true,
       upsert: true,
-      includeResultMetadata: true,
     },
   );
 
   return {
-    value: result.value,
-    updateExisting: result.lastErrorObject.updatedExisting,
+    value: result,
+    updateExisting: !!result,
   };
 };
